Move JCacheGaugeSet to the right directory
