+++ fullstack/src/main/java/org/terracotta/demo/config/jcache/JCacheGaugeSet.java
@@ -1,4 +1,4 @@
-package org.terracotta.demo.config;
+package org.terracotta.demo.config.jcache;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
